requirements.txt
pygame-ce
PySimpleGUI-4-foss